created entities for order items and inventory, created migrations

+++ src/main/resources/db/migration/V3__create_Inventory_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE inventory (
     id BIGINT PRIMARY KEY AUTO_INCREMENT,
-    product_id INT(11) NOT NULL,
+    product_id BIGINT NOT NULL,
     quantity INT(11) NOT NULL DEFAULT 0,
 
     CONSTRAINT fk_inventory_product
